+++ src/superface-test.interfaces.ts
@@ -49,4 +49,5 @@ export interface RecordingProcessOptions {
   processRecordings?: boolean;
   beforeRecordingSave?: ProcessingFunction;
   beforeRecordingLoad?: ProcessingFunction;
+  hideInput?: string[];
 }